Non-matching decomp of func_80083A58_53F08

+++ include/functions.us.h
@@ -181,7 +181,7 @@ void func_80083DBC_5426C(s16, u8);
 u8 func_8008DC34_5E0E4(s16, s16, s16, s16, s16);
 void func_8008EB20_5EFD0(s16, s16, s16, u16, u8, s32, s32);
 void func_800840C4_54574(u8);
-u8 func_80083A58_53F08(s32);
+u8 func_80083A58_53F08(u8);
 void func_8008CC3C_5D0EC(void);
 void func_8008DFA0_5E450(void);
 void func_8008EEC4_5F374(void);
